+++ Server/Server/Data/MetaData.cs
@@ -185,7 +185,7 @@ namespace Server.Data
         public int BaseDEX;
         public int BaseINT;
         public int BaseLUK;
-        public float CriticalRate;
+        public int CriticalRate;
         public float CriticalDamage;
     }
 }
